pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rmCalendarMacOS"
version = "0.1.0"
description = "Sync macOS Calendar to reMarkable tablets as interactive PDF planners"
# readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "icalendar>=6.0",
    "recurring-ical-events>=3.0",
    "httpx>=0.27",
    "reportlab>=4.0",
    "paramiko>=3.4",
    "Pillow>=10.0",
    "click>=8.1",
    "python-dateutil>=2.9",
    "textual>=1.0",
    "pyobjc-framework-EventKit>=10.0; sys_platform == 'darwin'",
]

[project.scripts]
rmcal = "rmcal.__main__:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pypdf>=4.0",
    "pytest-mock>=3.14",
]

[tool.hatch.build.targets.wheel]
packages = ["src/rmcal"]

[tool.pytest.ini_options]
testpaths = ["tests"]
